+++ kadu-core/chat/style-engines/chat-engine-kadu/chat-engine-kadu.cpp
@@ -171,7 +171,7 @@ QString KaduChatStyleEngine::formatMessage(MessageRenderInfo *message, MessageRe
 
 			includeHeader =
 				(aft.type() == MessageTypeSystem) ||
-				((msg.receiveDate().toTime_t() - aft.receiveDate().toTime_t() > (ChatStylesManager::instance()->cfgNoHeaderInterval() * 60)) ||
+				((static_cast<int>(msg.receiveDate().toTime_t() - aft.receiveDate().toTime_t()) > (ChatStylesManager::instance()->cfgNoHeaderInterval() * 60)) ||
 				 (msg.messageSender() != aft.messageSender()));
 		}
