+++ kadu-core/debug.cpp
@@ -45,7 +45,7 @@ bool KADUAPI showTimesInDebug = false;
 
 void _kdebug_with_mask(int mask, const char* file, const int line, const char* format,...)
 {
-	static int last;
+	static long last;
 	static QMutex mutex;
 
 	if (debug_mask & mask)
@@ -55,10 +55,10 @@ void _kdebug_with_mask(int mask, const char* file, const int line, const char* f
 		if (showTimesInDebug)
 		{
 			time_t sec;
-			int msec;
+			long int msec;
 			getTime(&sec, &msec);
-			int x = (sec % 1000) * 1000 + msec;
-			fprintf(stderr, "KK <%d:%03d:%09d:%s:%i>\t", (int)sec, msec, x - last, file, line);
+			long x = (sec % 1000) * 1000 + msec;
+			fprintf(stderr, "KK <%d:%03ld:%09ld:%s:%i>\t", (int)sec, msec, x - last, file, line);
 			last = x;
 		}
 		else
