+++ kadu-core/gui/widgets/configuration/config-list-widget.cpp
@@ -118,10 +118,10 @@ void ConfigListWidget::setVisible(bool visible)
 bool ConfigListWidget::fromDomElement(QDomElement domElement)
 {
 	QDomNodeList children = domElement.childNodes();
-	int length = children.length();
-	for (int i = 0; i < length; i++)
+	uint length = children.length();
+	for (uint i = 0; i < length; i++)
 	{
-		QDomNode node = children.item(i);
+		QDomNode node = children.item(static_cast<int>(i));
 		if (node.isElement())
 		{
 			QDomElement element = node.toElement();
