+++ kadu-core/parser/parser.cpp
@@ -669,7 +669,7 @@ QString Parser::parse(const QString &s, Talkable talkable, const QObject * const
 									tokens.at(firstSpaceTokenIdx).rawContent().left(spacePos);
 
 						if (filePath.startsWith(QLatin1String("file://")))
-							filePath = filePath.mid(qstrlen("file://"));
+							filePath = filePath.mid(static_cast<int>(qstrlen("file://")));
 
 						bool checkFileExists = (pe2.type() == PT_CHECK_FILE_EXISTS);
 						if (QFile::exists(filePath) == checkFileExists)
